+++ src/rawsock-getroute.c
@@ -441,7 +441,8 @@ again:
         goto again;
     }
     if (err != NO_ERROR) {
-        fprintf(stderr, "GetAdaptersInfo failed with error: %u\n", err);
+        fprintf(stderr, "GetAdaptersInfo failed with error: %u\n", 
+                            (unsigned)err);
         return EFAULT;
     }
