+++ src/rawsock.c
@@ -86,13 +86,16 @@ int pcap_setdirection(pcap_t *pcap, pcap_direction_t direction)
     if (real_setdirection == 0) {
         HMODULE h = LoadLibraryA("wpcap.dll");
         if (h == NULL) {
-            fprintf(stderr, "couldn't load wpcap.dll: %u\n", GetLastError());
+            fprintf(stderr, "couldn't load wpcap.dll: %u\n", 
+                                (unsigned)GetLastError());
             return -1;
         }
 
-        real_setdirection = (int (*)(pcap_t*,pcap_direction_t))GetProcAddress(h, "pcap_setdirection");
+        real_setdirection = (int (*)(pcap_t*,pcap_direction_t))
+                            GetProcAddress(h, "pcap_setdirection");
         if (real_setdirection == 0) {
-            fprintf(stderr, "couldn't find pcap_setdirection(): %u\n", GetLastError());
+            fprintf(stderr, "couldn't find pcap_setdirection(): %u\n", 
+                                (unsigned)GetLastError());
             return -1;
         }
     }
@@ -195,7 +198,8 @@ rawsock_init(void)
 
         }
     } else {
-        printf("GetAdaptersInfo failed with error: %d\n", dwRetVal);
+        printf("GetAdaptersInfo failed with error: %u\n", 
+                                                    (unsigned)dwRetVal);
 
     }
     if (pAdapterInfo)
@@ -356,6 +360,7 @@ int rawsock_recv_packet(
     unsigned *usecs,
     const unsigned char **packet)
 {
+        /*--*/LOG(1, "%u 0x%llx\n", __LINE__, (unsigned long long)adapter->ring);
     if (adapter->ring) {
         struct pfring_pkthdr hdr;
         int err;
@@ -412,17 +417,20 @@ rawsock_send_probe(
     unsigned seqno, unsigned flush,
     struct TemplateSet *tmplset)
 {
+        /*--*/LOG(1, "%u 0x%llx\n", __LINE__, (unsigned long long)adapter->ring);
     /*
      * Construct the destination packet
      */
     template_set_target(tmplset, ip_them, port_them, ip_me, port_me, seqno);
     if (tmplset->length < 60)
         tmplset->length = 60;
+        /*--*/LOG(1, "%u 0x%llx\n", __LINE__, (unsigned long long)adapter->ring);
 
     /*
      * Send it
      */
     rawsock_send_packet(adapter, tmplset->px, tmplset->length, flush);
+        /*--*/LOG(1, "%u 0x%llx\n", __LINE__, (unsigned long long)adapter->ring);
 }
 
 
@@ -631,6 +639,7 @@ rawsock_init_adapter(const char *adapter_name,
 
     if (is_offline)
         return adapter;
+        /*--*/LOG(1, "%u 0x%llx\n", __LINE__, (unsigned long long)adapter->ring);
 
     /*----------------------------------------------------------------
      * PORTABILITY: WINDOWS
@@ -647,6 +656,7 @@ rawsock_init_adapter(const char *adapter_name,
         } else
             adapter_name = new_adapter_name;
     }
+        /*--*/LOG(1, "%u 0x%llx\n", __LINE__, (unsigned long long)adapter->ring);
 
     /*----------------------------------------------------------------
      * PORTABILITY: PF_RING
@@ -717,6 +727,8 @@ rawsock_init_adapter(const char *adapter_name,
         return adapter;
     }
 
+        /*--*/LOG(1, "%u 0x%llx\n", __LINE__, (unsigned long long)adapter->ring);
+    
     /*----------------------------------------------------------------
      * PORTABILITY: LIBPCAP
      *
@@ -725,12 +737,14 @@ rawsock_init_adapter(const char *adapter_name,
     {
         LOG(1, "pcap: %s\n", pcap_lib_version());
         LOG(2, "pcap:'%s': opening...\n", adapter_name);
+        /*--*/LOG(1, "%u 0x%llx\n", __LINE__, (unsigned long long)adapter->ring);
         adapter->pcap = pcap_open_live(
                     adapter_name,           /* interface name */
                     65536,                  /* max packet size */
                     8,                      /* promiscuous mode */
                     1000,                   /* read timeout in milliseconds */
                     errbuf);
+        /*--*/LOG(1, "%u 0x%llx\n", __LINE__, (unsigned long long)adapter->ring);
         if (adapter->pcap == NULL) {
             LOG(0, "FAIL: %s\n", errbuf);
             if (strstr(errbuf, "perm")) {
@@ -747,6 +761,7 @@ rawsock_init_adapter(const char *adapter_name,
             return 0;
         } else
             LOG(1, "pcap:'%s': successfully opened\n", adapter_name);
+        /*--*/LOG(1, "%u 0x%llx\n", __LINE__, (unsigned long long)adapter->ring);
 
         /* Figure out the link-type. We suport Ethernet and IP */
         {
@@ -765,6 +780,7 @@ rawsock_init_adapter(const char *adapter_name,
 
             }
         }
+        /*--*/LOG(1, "%u 0x%llx\n", __LINE__, (unsigned long long)adapter->ring);
 
         /* Set any BPF filters the user might've set */
         if (bpf_filter) {
@@ -789,6 +805,9 @@ rawsock_init_adapter(const char *adapter_name,
                 exit(1);
             }
         }
+
+        /*--*/LOG(1, "%u 0x%llx\n", __LINE__, (unsigned long long)adapter->ring);
+
     }
 
     /*----------------------------------------------------------------
